requirements.txt
django
djangorestframework
django-cors-headers
feedparser
requests
python-dateutil
gunicorn
